fix: resolve CI-only warnings-as-errors failures
Two issues surfaced by CI's macOS/Ubuntu matrix that this session's
local (Apple Clang, older diagnostic set) warnings-as-errors build did
not reproduce:

- include/espressolab/cfd.hpp used std::shared_ptr without including
  <memory>. This compiled locally because libc++ pulls it in
  transitively via another header; GCC/libstdc++ on Ubuntu CI does
  not. Pre-existing latent bug, previously masked on Ubuntu CI by an
  earlier -Werror=shadow failure in simulator.cpp that aborted the
  build before reaching this translation unit (that shadow issue is
  already fixed on this branch by the puck-region refactor in
  8a9badc, which is what let CI get far enough to expose this one).
- apps/espressolab_cli/workflows.cpp:180 (bench p95 calculation)
  implicitly converted samples_ms.size() from std::size_t to double
  inside `0.95 * (...)`. Apple Clang on the GitHub macOS runner flags
  this under -Wconversion; this session's local Clang did not. Made
  the conversion explicit.

Verified locally: full build, 146/146 tests, demo determinism check,
10/10 PTY smoke checks all still pass. Could not reproduce the Ubuntu
GCC path locally (no GCC toolchain on this machine) or the exact Apple
Clang diagnostic locally (older Xcode); relying on CI to confirm both.

+++ apps/espressolab_cli/workflows.cpp
@@ -177,7 +177,8 @@ BenchOutcome run_bench(const BenchRequest& request, const CancellationCallback& 
     std::sort(outcome.samples_ms.begin(), outcome.samples_ms.end());
     outcome.best_ms = outcome.samples_ms.front();
     outcome.median_ms = outcome.samples_ms[outcome.samples_ms.size() / 2];
-    outcome.p95_ms = outcome.samples_ms[static_cast<std::size_t>(0.95 * (outcome.samples_ms.size() - 1))];
+    outcome.p95_ms = outcome.samples_ms[static_cast<std::size_t>(
+        0.95 * static_cast<double>(outcome.samples_ms.size() - 1))];
     return outcome;
 }
